+++ client/factory.go
@@ -18,6 +18,7 @@ import (
 	"github.com/thanos-io/objstore/providers/oci"
 	"github.com/thanos-io/objstore/providers/oss"
 	"github.com/thanos-io/objstore/providers/s3"
+	"github.com/thanos-io/objstore/providers/storj"
 	"github.com/thanos-io/objstore/providers/swift"
 
 	"github.com/go-kit/log"
@@ -39,6 +40,7 @@ const (
 	ALIYUNOSS  ObjProvider = "ALIYUNOSS"
 	BOS        ObjProvider = "BOS"
 	OCI        ObjProvider = "OCI"
+	STORJ      ObjProvider = "STORJ"
 	OBS        ObjProvider = "OBS"
 )
 
@@ -82,6 +84,8 @@ func NewBucket(logger log.Logger, confContentYaml []byte, reg prometheus.Registe
 		bucket, err = bos.NewBucket(logger, config, component)
 	case string(OCI):
 		bucket, err = oci.NewBucket(logger, config)
+	case string(STORJ):
+		bucket, err = storj.NewBucket(logger, config, component)
 	case string(OBS):
 		bucket, err = obs.NewBucket(logger, config)
 	default: